+++ JTSim/Simulator/Systems/Specific/AR.cs
@@ -10,18 +10,18 @@ namespace JTSim
         równanie różnicowe: y(i) = 0.5y(i-1) + 0.2y(i-2) + 0.2y(i-3) + 0.5u(i-2) + u(i-3)
         */
 
-        public AR (float y_1, float y_2, float y_3)
+        public AR (double y_1, double y_2, double y_3)
         {
-            initStates = Vector<float>.Build.Dense(new float[] { y_1, y_2, y_3 });
-            initInputs = Vector<float>.Build.Dense(3, 0f);
+            initStates = Vector<double>.Build.Dense(new double[] { y_1, y_2, y_3 });
+            initInputs = Vector<double>.Build.Dense(3, 0d);
         }
 
-        public override float DifferenceEquasion(Vector<float> states, Vector<float> inputs, float t, float h)
+        public override double DifferenceEquasion(Vector<double> states, Vector<double> inputs, double t, double h)
         {
-            return 0.5f * states[0] + 0.2f * states[1] + 0.2f * states[2] + 0.5f * inputs[1] + 1f * inputs[2];
+            return 0.5d * states[0] + 0.2d * states[1] + 0.2d * states[2] + 0.5d * inputs[1] + 1d * inputs[2];
         }
 
-        public override float OutputEquation(Vector<float> states, Vector<float> inputs)
+        public override double OutputEquation(Vector<double> states, Vector<double> inputs)
         {
             return states[0];
         }
