+++ JTSim/Simulator/Systems/Specific/AlphaFilter.cs
@@ -4,25 +4,25 @@ namespace JTSim
 {
     public class AlphaFilter : GenericSystem
     {
-        float k = 1f;
-        float T = 1f;
+        double k = 1d;
+        double T = 1d;
 
-        float initState;
+        double initState;
 
-        public AlphaFilter (float initState)
+        public AlphaFilter (double initState)
         {
             this.initState = initState;
         }
 
-        public override void Init(float h)
+        public override void Init(double h)
         {
             output = initState;
         }
 
-        public override void Step(float u, float t, float h)
+        public override void Step(double u, double t, double h)
         {
-            float alpha = (float)Math.Exp(-h / T);
-            output = output * alpha + (1f - alpha) * u * k;
+            double alpha = (double)Math.Exp(-h / T);
+            output = output * alpha + (1d - alpha) * u * k;
         }
     }
 }
